Initial Commit
This project has been halted for a while, and was written in suboptimal
conditions and ways, that now I'm having hard time to remember what I
was thinking when I wrote, what I wrote. Thats because I'm in a busy
point in my life, I couldnt spare time to look at this little project
since forever. Looks like the init commit has also been forgotten in the
process.

This commit aims to get a snapshot of the project as it is, and try to
to understand, and continue working on it. If things go sour, I might
as well consider scraping it and starting from anew.